Major refactor/cleanup of the Environment/Device APIs.

+++ jota/jota-core/src/main/java/com/qxotic/jota/runtime/KernelArgs.java
@@ -122,7 +122,7 @@ public final class KernelArgs {
             return (int) bits;
         }
         if (type == DataType.I64) {
-            return (long) bits;
+            return bits;
         }
         if (type == DataType.FP16) {
             return Float.float16ToFloat((short) bits);
@@ -159,7 +159,7 @@ public final class KernelArgs {
             return (int) bits;
         }
         if (type == DataType.I64) {
-            return (long) bits;
+            return bits;
         }
         if (type == DataType.FP16) {
             return Float.float16ToFloat((short) bits);
